Login now redirects jwt to /login
