+++ containers/vuejs/src/components/HomePage.vue
@@ -3,20 +3,6 @@ import Chat from './Chat.vue'
 import GameHeader from './GameHeader.vue'
 import PongCanvas from './PongCanvas.vue'
 import Sidebar from './Sidebar.vue'
-import { useRouter } from 'vue-router'
-
-function extractJwtFromUrl() {
-  const urlParams = new URLSearchParams(window.location.search)
-  const jwt = urlParams.get('jwt')
-  if (jwt) {
-    localStorage.setItem('jwt', jwt)
-  }
-  router.replace({ path: '/' }) // Trims the jwt parameter from the URL
-}
-
-const router = useRouter()
-
-extractJwtFromUrl()
 </script>
 
 <template>
