+++ containers/vuejs/src/components/PongCanvas.vue
@@ -7,8 +7,8 @@
 
 <script setup lang="ts">
 import { ref, onMounted } from 'vue'
-import { gameSocket } from './SocketManager.ts'
 import ScoreBoard from './ScoreBoard.vue'
+import { gameSocket } from './getSocket'
 
 gameSocket.on('pong', (data: any) => {
   render(data)
