+++ containers/vuejs/src/components/ScoreBoard.vue
@@ -8,7 +8,8 @@
 
 <script setup lang="ts">
 import { ref } from 'vue'
-import { gameSocket } from './SocketManager.ts'
+import { gameSocket } from './getSocket'
+
 const leftPlayerScore = ref(0)
 const rightPlayerScore = ref(0)
 const startGame = ref(false)
